:fire: template.py is gone
