+++ pyecharts/__init__.py
@@ -37,6 +37,6 @@ from pyecharts.custom.page import Page
 from pyecharts.custom.timeline import Timeline
 
 # misc
-from pyecharts.template import online
-from pyecharts.template import configure
+from pyecharts.conf import online
+from pyecharts.conf import configure
 from pyecharts.style import Style
